Factor out some modules from storage
